Make code modular.
